IGNITE-25169 Split BinarySchemaBuilder from BinarySchema
