+++ modules/core/src/main/java/org/apache/ignite/internal/binary/BinarySchemaBuilder.java
@@ -1,0 +1,62 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.ignite.internal.binary;
+
+import java.util.ArrayList;
+
+/**
+ * Binary schema builder.
+ */
+class BinarySchemaBuilder {
+    /**
+     * Schema ID.
+     */
+    private int schemaId = BinaryUtils.schemaInitialId();
+
+    /**
+     * Fields.
+     */
+    private final ArrayList<Integer> fields = new ArrayList<>();
+
+    /**
+     *
+     */
+    BinarySchemaBuilder() {
+        // No-op.
+    }
+
+    /**
+     * Add field.
+     *
+     * @param fieldId Field ID.
+     */
+    public void addField(int fieldId) {
+        fields.add(fieldId);
+
+        schemaId = BinaryUtils.updateSchemaId(schemaId, fieldId);
+    }
+
+    /**
+     * Build schema.
+     *
+     * @return Schema.
+     */
+    public BinarySchema build() {
+        return new BinarySchema(schemaId, fields);
+    }
+}
